+++ example/src/views/Komponenter/TjekboksExample.vue
@@ -1,21 +1,21 @@
 <template>
   <section>
-    <xfds-formgroup label="Vælg en checkbox">
+    <fds-formgroup label="Vælg en checkbox">
       <fds-checkbox-list v-model="checkboxList">
         <template v-slot:[`melon`]>
           <p>Det er muligt at benytte radio til mere indhold</p>
         </template>
       </fds-checkbox-list>
-    </xfds-formgroup>
+    </fds-formgroup>
 
     <fds-pre header="v-model" :json="checkboxList" />
 
-    <xfds-formgroup label="Single Checkbox">
+    <fds-formgroup label="Single Checkbox">
       <fds-checkbox v-model="oneChecked" class="mt-2"> Et valg </fds-checkbox>
       <fds-checkbox v-model="twoChecked" small> Andet valg - small </fds-checkbox>
 
       <fds-pre header="v-model" :json="{ oneChecked, twoChecked }" />
-    </xfds-formgroup>
+    </fds-formgroup>
   </section>
 </template>