+++ src/components/extra/xfds-form-radio.vue
@@ -5,7 +5,7 @@
     #default="{ isValid, errorMessage }"
     :dirty="dirty"
     @valid="validEvent">
-    <xfds-formgroup
+    <fds-formgroup
       :is-valid="isValid"
       :label="label"
       :tooltip="tooltip"
@@ -22,7 +22,7 @@
         @dirty="touchedEvent">
         <slot />
       </fds-radio>
-    </xfds-formgroup>
+    </fds-formgroup>
   </xfds-validate>
 </template>